chore: rename folder headline to headlineSection
