+++ src/components/landing-page/main/headlineSection/headphoneIllustration.tsx
